feat: add admin/partner order controls with pagination and global error handling
Implemented paginated order APIs, separated Admin/Partner controller logic, and integrated a global exception handling mechanism for robust error management and security.

+++ src/main/java/com/spareparts/spareparts_backend/dto/OrderItemDto.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,6 +15,6 @@ public class OrderItemDto {
     private Integer spareItemId;
     private String spareItemName;
     private int quantity;
-    private double unitPrice;
-    private double totalPrice;
+    private BigDecimal unitPrice;
+    private BigDecimal totalPrice;
 }
